==> 5_predicting_popularity_based_on_attributes.ipynb <==
### Import packages and data

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
import warnings
warnings.filterwarnings("ignore")

df = pd.read_csv('data.csv')
pd.set_option('display.max_columns', None)
df.head()

## Preparing data for models

df.columns

from sklearn.model_selection import train_test_split
df = df[['valence', 'year', 'acousticness', 'artists', 'danceability','duration_ms', 'energy', 'explicit', 'id', 'instrumentalness', 'key','liveness', 'loudness', 'mode', 'name', 'release_date','speechiness', 'tempo', 'popularity']].dropna()
X = df[['acousticness', 'danceability', 'duration_ms', 'energy','instrumentalness', 'liveness', 'loudness', 'speechiness', 'tempo', 'valence', 'key', 'mode']]
y = df['popularity']
X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=126)

from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

from sklearn.preprocessing import PolynomialFeatures
pf = PolynomialFeatures()
X_train_expanded = pf.fit_transform(X_train)
X_test_expanded = pf.transform(X_test)

## Dummy (Baseline)

import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)

from sklearn.dummy import DummyRegressor
dr = DummyRegressor()
dr.fit(X_train, y_train)
print('R-squared score for dummy model with raw values: ', dr.score(X_test, y_test))

## Linear Regression

### Raw Values

from sklearn.linear_model import LinearRegression
lr = LinearRegression()
lr.fit(X_train, y_train)
print('R-squared score for linear regression with raw values: ', lr.score(X_test, y_test))

### Scaled Values

from sklearn.linear_model import LinearRegression
lr = LinearRegression()
lr.fit(X_train_scaled, y_train)
print('R-squared score for linear regression with scaled values: ', lr.score(X_test_scaled, y_test))

### Polynomial Features

from sklearn.linear_model import LinearRegression
lr = LinearRegression()
lr.fit(X_train_expanded, y_train)
print('R-squared score for linear regression with polynomial values: ', lr.score(X_test_expanded, y_test))

## Ridge

### Raw Values

from sklearn.linear_model import Ridge
ridge = Ridge(alpha=1).fit(X_train, y_train)
print('R-squared score for ridge regression with raw values: ', ridge.score(X_test, y_test))

from sklearn.model_selection import GridSearchCV
params = {'alpha': [0.001, 0.01, 0.1, 1, 10, 100, 1000]}
gs = GridSearchCV(Ridge(), param_grid=params, scoring='r2', n_jobs=-1)
gs.fit(X_train, y_train)
print('Optimal alpha for ridge regression with raw values: ', gs.best_params_)
print('R-squared score for ridge regression with raw values: ', gs.score(X_test, y_test))

### Scaled Values

ridge = Ridge(alpha=1).fit(X_train_scaled, y_train)
print('R-squared score for ridge regression with scaled values: ', ridge.score(X_test_scaled, y_test))

from sklearn.model_selection import GridSearchCV
params = {'alpha': [0.001, 0.01, 0.1, 1, 10, 100, 1000]}
gs = GridSearchCV(Ridge(), param_grid=params, scoring='r2', n_jobs=-1)
gs.fit(X_train_scaled, y_train)
print('Optimal alpha for ridge regression with scaled values: ', gs.best_params_)
print('R-squared score for ridge regression with scaled values: ', gs.score(X_test_scaled, y_test))

### Polynomial Features

ridge = Ridge(alpha=1).fit(X_train_expanded, y_train)
print('R-squared score for ridge regression with expanded values: ', ridge.score(X_test_expanded, y_test))

from sklearn.model_selection import GridSearchCV
params = {'alpha': [0.001, 0.01, 0.1, 1, 10, 100, 1000]}
gs = GridSearchCV(Ridge(), param_grid=params, scoring='r2', n_jobs=-1)
gs.fit(X_train_expanded, y_train)
print('Optimal alpha for ridge regression with expanded values: ', gs.best_params_)
print('R-squared score for ridge regression with expanded values: ', gs.score(X_test_expanded, y_test))

## Lasso

### Raw

from sklearn.linear_model import Lasso
lasso = Lasso(alpha=1).fit(X_train, y_train)
print('R-squared score for lasso regression with raw values: ', lasso.score(X_test, y_test))

params = {'alpha': [0.001, 0.01, 0.1, 1, 10, 100, 1000]}
gs = GridSearchCV(estimator=lasso, param_grid=params, scoring='r2', n_jobs=-1)
gs.fit(X_train, y_train)
print('Optimal alpha for lasso regression with raw values: ', gs.best_params_)
print('R-squared score for lasso regression with raw values: ', gs.score(X_test, y_test))

### Scaled

lasso = Lasso(alpha=1).fit(X_train_scaled, y_train)
print('R-squared score for lasso regression with scaled values: ', lasso.score(X_test_scaled, y_test))

params = {'alpha': [0.001, 0.01, 0.1, 1, 10, 100, 1000]}
gs = GridSearchCV(estimator=lasso, param_grid=params, scoring='r2', n_jobs=-1)
gs.fit(X_train_scaled, y_train)
print('Optimal alpha for lasso regression with scaled values: ', gs.best_params_)
print('R-squared score for lasso regression with scaled values: ', gs.score(X_test_scaled, y_test))

### Polynomial features

from sklearn.linear_model import Lasso
lasso = Lasso(alpha=1).fit(X_train_expanded, y_train)
print('R-squared score for lasso regression with expanded values: ', lasso.score(X_test_expanded, y_test))

params = {'alpha': [0.001, 0.01, 0.1, 1, 10, 100, 1000]}
gs = GridSearchCV(estimator=lasso, param_grid=params, scoring='r2', n_jobs=-1)
gs.fit(X_train_expanded, y_train)
print('Optimal alpha for lasso regression with expanded values: ', gs.best_params_)
print('R-squared score for lasso regression with expanded values: ', gs.score(X_test_expanded, y_test))

## ElasticNet

### Raw

from sklearn.linear_model import ElasticNet
elastic_net = ElasticNet(alpha=1, l1_ratio=0.5).fit(X_train, y_train)
print('R-squared score for ElasticNet regression with raw values: ', elastic_net.score(X_test, y_test))

alpha = [0.001, 0.01, 0.1, 1, 10, 100, 1000]
l1_ratio = [0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1]
param_grid = dict(alpha=alpha, l1_ratio=l1_ratio)
gs = GridSearchCV(estimator=elastic_net, param_grid=param_grid, scoring='r2', n_jobs=-1)
gs.fit(X_train, y_train)
print('Optimal alpha and L1 ratio for ElasticNet regression with raw values: ', gs.best_params_)
print('R-squared score for ElasticNet regression with raw values: ', gs.score(X_test, y_test))

### Scaled

from sklearn.linear_model import ElasticNet
elastic_net = ElasticNet(alpha=1, l1_ratio=0.5).fit(X_train_scaled, y_train)
print('R-squared score for ElasticNet regression with scaled values: ', elastic_net.score(X_test_scaled, y_test))

alpha = [0.001, 0.01, 0.1, 1, 10, 100, 1000]
l1_ratio = [0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1]
param_grid = dict(alpha=alpha, l1_ratio=l1_ratio)
gs = GridSearchCV(estimator=elastic_net, param_grid=param_grid, scoring='r2', n_jobs=-1)
gs.fit(X_train_scaled, y_train)
print('Optimal alpha and L1 ratio for ElasticNet regression with scaled values: ', gs.best_params_)
print('R-squared score for ElasticNet regression with scaled values: ', gs.score(X_test_scaled, y_test))

### Polynomial features

from sklearn.linear_model import ElasticNet
elastic_net = ElasticNet(alpha=1, l1_ratio=0.5).fit(X_train_expanded, y_train)
print('R-squared score for ElasticNet regression with expanded values: ', elastic_net.score(X_test_expanded, y_test))

alpha = [0.001, 0.01, 0.1, 1, 10, 100, 1000]
l1_ratio = [0, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1]
param_grid = dict(alpha=alpha, l1_ratio=l1_ratio)
gs = GridSearchCV(estimator=elastic_net, param_grid=param_grid, scoring='r2', n_jobs=-1)
gs.fit(X_train_expanded, y_train)
print('Optimal alpha and L1 ratio for ElasticNet regression with expanded values: ', gs.best_params_)
print('R-squared score for ElasticNet regression with expanded values: ', gs.score(X_test_expanded, y_test))
