==> 6_get_grammy_data.ipynb <==
# This section is a Work in Progress

# https://developer.spotify.com/documentation/web-api/reference/#endpoint-get-audio-features

import pandas as pd
import requests
from bs4 import BeautifulSoup
import plotly.express as px
import urllib
import os

get_nominees = False

if get_nominees == True:
    r = requests.get('https://en.wikipedia.org/wiki/Grammy_Award_for_Song_of_the_Year')
    r_html = r.text
    soup = BeautifulSoup(r_html)
    noms_lists = soup.find_all('ul')[2:-53]
    noms_lists = noms_lists[0:20] + noms_lists[19:]
    nominees = pd.read_html(r_html)[-4]
    nominees['noms_lists'] = noms_lists
    nominees = nominees[['Year[I]', 'Work','Performing artist(s)[II]', 'Nominees', 'noms_lists']]
    nominees.rename(columns={'Year[I]': 'year', 'Work': 'song','Performing artist(s)[II]': 'musician', 'Nominees': 'nominee_raw_list', 'noms_lists': 'nominees_html'}, inplace=True)
    df_nominees = pd.DataFrame(columns=['year', 'nominee'])
    def get_noms(row):
        year_html = row['nominees_html'].find_all('li')
        for item in year_html:
            try:
                title_arr = item.text.split('for "')[1].replace('"', '')
            except:
                title_arr = item.text.replace('"', '')
            df_nominees.loc[len(df_nominees.index)] = [row['year'], title_arr]
    nominees.apply(get_noms, axis=1)
    df_nominees['won_award'] = 0
    df_winners = nominees[['year', 'song', 'musician']]
    df_winners['won_award'] = 1
    df_nominees['song'] = df_nominees['nominee'].apply(lambda x: x.rsplit(' performed by ', 1)[0])
    df_nominees['musician'] = df_nominees['nominee'].apply(lambda x: x.rsplit(' performed by ', 1)[1])
    df_nominees = df_nominees[['year', 'song', 'musician', 'won_award']]
    df_all = pd.concat([df_winners, df_nominees], axis=0)
    df_all.sort_values(by='year', inplace=True)
    df_all['search_term'] = df_all['song'] + ' ' + df_all['musician']
    df_all.dropna(subset=['song'], inplace=True)
    df_all.reset_index(drop=True, inplace=True)
    df_all.to_csv(r'grammy_winners_and_nominees.csv')
else:
    df_all = pd.read_csv(r'grammy_winners_and_nominees.csv')

df_all = df_all[df_all['ids'] != '000']
df_all.reset_index(drop=True, inplace=True)
df_all

import spotipy
from spotipy.oauth2 import SpotifyClientCredentials

sp = spotipy.Spotify(auth_manager=SpotifyClientCredentials(client_id=os.environ["SPOTIPY_CLIENT_ID"], client_secret=os.environ["SPOTIPY_CLIENT_SECRET"]))
# ids = []
for term in df_all['search_term'][326:]:
    print('GETTING API RESULTS - INTERRUPT KERNEL IF THIS IS NOT INTENDED')
    results = sp.search(q=term.replace('"', '').replace('*', ''), limit=1)
    try:
        song_id = results['tracks']['items'][0]['id']
    except:
        song_id = '000'
    ids.append(song_id)
df_all['ids'] = ids
print('done')

df_all.reset_index(drop=True, inplace=True)

df_all

import time
get_audio_features = True
if get_audio_features:
    print('GETTING API RESULTS - INTERRUPT KERNEL IF THIS IS NOT INTENDED')
    import requests
    attributes = []
    for song_id in df_all['ids']:
        url = f'https://api.spotify.com/v1/audio-features/{song_id}'
        payload={}
        headers = {

        }
        response = requests.request("GET", url, headers=headers, data=payload)
        attributes.append(response.text)
        time.sleep(1)
    df_all['attributes'] = attributes
df_all = df_all[df_all['attributes'] != '']
df_all.reset_index(drop=True, inplace=True)

music_attrs = pd.DataFrame(columns=['danceability', 'energy', 'key', 'loudness', 'mode', 'speechiness', 'acousticness', 'instrumentalness', 'liveness', 'valence', 'tempo', 'type', 'id', 'uri', 'track_href', 'analysis_url', 'duration_ms', 'time_signature'])
def get_attrs(row):
    music_attrs.loc[len(music_attrs.index)] = [json.loads(row)[key] for key in json.loads(row).keys()]
df_all['attributes'].apply(get_attrs)

df_all_v2 = pd.merge(df_all, music_attrs, how='left', left_on='ids', right_on='id')
df_all_v2 = df_all_v2.drop_duplicates().reset_index(drop=True).drop(columns=['attributes', 'ids'])
df_all_v2 = df_all_v2[['year', 'song', 'musician', 'danceability','energy', 'key', 'loudness', 'mode', 'speechiness', 'acousticness','instrumentalness', 'liveness', 'valence', 'tempo', 'duration_ms', 'time_signature',  'uri', 'track_href', 'analysis_url', 'won_award']]
df_all_v2.to_csv(r'grammy_nominees_with_music_attributes.csv')
df_historical = df_all_v2[df_all_v2['year'] != 2021]
df_2021_nominees = df_all_v2[df_all_v2['year'] == 2021]

df_historical

df_2021_nominees
